front/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
fastapi==0.95.1
ffmpeg-python==0.2.0
filelock==3.12.0
frozenlist==1.3.3
future==0.18.3
h11==0.14.0
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
langchain==0.0.168
llvmlite==0.40.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.1.1
networkx==3.1
numba==0.57.0
numexpr==2.8.4
numpy==1.24.2
openai==0.27.4
openai-whisper==20230314
openapi-schema-pydantic==1.2.4
packaging==23.1
peewee==3.16.2
psycopg2-binary==2.9.6
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starlette==0.26.1
sympy==1.12
tenacity==8.2.2
tiktoken==0.3.1
torch==2.0.1
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
